requirements.txt
crewai
fastapi
uvicorn
langchain-openai
python-dotenv
